Heuristik-Variante erreicht jetzt auch den aufzeichnenden Tiling-Pfad
Der Durchreich-Commit von heute war nicht fertig, und die Abnahme hat es nicht
gemerkt. Die Variante kam bis in die PlayerLoopConfig, aber der Verbraucher
sass nur im Arena-Zweig: sobald aufgezeichnet wird -- also auf JEDEM
Self-Play-Pfad -- lief die varianten-blinde Fassung von tiling_step. Ein
200-Partien-Pilot mit v2huelle war deshalb bis auf die letzte Nachkommastelle
identisch mit dem v1-Kontrollarm.

Aufgefallen ist es an genau der Signatur, die dafuer in der Fallenliste steht:
gleiche Seeds, gleiche Zahlen. Nicht an der Abnahme, die den Fall pruefen
sollte -- die hatte auf einem Arena-Pfad gemessen, wo der Verbraucher sitzt,
und damit den einzigen Pfad verfehlt, fuer den der Umbau gebaut war.

Die zweite Haelfte fehlt weiterhin und ist hier NICHT mitgebaut: die
Draft-Entscheidung. HeuristicArenaAgent traegt die Variante und wendet den
v2-Vorzug an, HeuristicSelfPlayAgent hat gar kein Variantenfeld. Ein Korpus
aus diesem Stand ist ein v1-Korpus mit leicht anderem Tiling-Routing, kein
v2huelle-Korpus. Das ist ein eigener Zuschnitt, weil der Self-Play-Agent
anders als der Arena-Agent eine POLICY-Verteilung liefern muss -- ein harter
Vorzug wuerde das Ziel auf one-hot zusammenfalten.

Dazu zwei Werkzeuge: der Long-dtype-Absturz in diagnosis.py Punkt 1, und
corpus_sanity_check.py fuer Reihen, Spalten und Plattenpunkte je Kriterium
(diagnosis.py deckt Policy, Strafleiste und Ergebnis ab, diese drei nicht).

Suite 526/0.

+++ tools/diagnosis.py
@@ -45,7 +45,7 @@ def run_diagnosis(data_dir: str, label: str):
 
     zero_mask   = (masks.sum(1) == 0).sum().item()
     leak        = (targets_p * (1 - masks)).sum(1).max().item()
-    mask_mean   = masks.sum(1).mean().item()
+    mask_mean   = masks.sum(1).float().mean().item()
     mask_min    = masks.sum(1).min().item()
     mask_max    = masks.sum(1).max().item()
